pyproject.toml
[project]
name = "cimig-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "wu",email = "wu@123.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest (>=9.0.2,<10.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "openai (>=2.28.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
